+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Unreleased
       which returns a `Result<&'static [ResourceType], StoreObjectConversionError>`
 - Add missing `StoreObject::Reactor` to the `seasonal-season-5` feature
 - Implement `Serialize` and `Deserialize` for `RoomStatus`
+- Add function `JsHashMap::entries`
 
 ### Bugfixes: